Slowly transition from ThermalFoundation -> Zairyou

+++ src/main/java/cofh/thermalexpansion/util/BehaviorFlorbDispense.java
@@ -23,7 +23,7 @@ public class BehaviorFlorbDispense extends BehaviorDefaultDispenseItem {
 
 		if (fluid != null) {
 			EntityFlorb florb = new EntityFlorb(world, pos.getX(), pos.getY(), pos.getZ(), fluid);
-			florb.shoot((double) facing.getFrontOffsetX(), (double) ((float) facing.getFrontOffsetY() + 0.1F), (double) facing.getFrontOffsetZ(), getProjectileVelocity(), getProjectileInaccuracy());
+			florb.shoot(facing.getFrontOffsetX(), (float) facing.getFrontOffsetY() + 0.1F, facing.getFrontOffsetZ(), getProjectileVelocity(), getProjectileInaccuracy());
 			world.spawnEntity(florb);
 			stack.shrink(1);
 		} else {
